refactor: replace useFontScaleTracking with useScreenReaderAndFontScaleTracking
